requirements.txt
streamlit
langgraph
langchain
python-docx
reportlab
PyMuPDF
transformers
torch
ollama-client
openai
python-dotenv
